requirements.txt
aiogram==3.8.0
aiosqlite==0.19.0
python-dotenv==1.0.0